Add direct render support
Allow libavcodec to decoder directly into VapourSynth's buffers
by implementing our own get_buffer and relese_buffer.

Since libavcodec requires a height that is a multiple of 32, and
a width that is a multiple of 16 for 4:2:0 YUV, we can only
actually use the direct-rendered buffer if our width and height
meet this requirement, since VapoutSynth has no way to allocate
a buffer larger than the width and height is given, and no way
to return a frame from pointers.

Should provide a non-trivial speedup on resolutions that meet this
requirement (e.g. 720x480).

Signed-off-by: Derek Buitenhuis <derek.buitenhuis@gmail.com>