Add 'nocrop' parameter
If this is set to true, direct rendering will always be used,
and no memcpy will take place, but it will retrurn the frame
with the aligned width and height, which may need cropping,
which constitutes a memcpy in vapoursynth.

If the image has to be cropped anyway, this function can provide
a speedup in the form of less memcpys.

Signed-off-by: Derek Buitenhuis <derek.buitenhuis@gmail.com>