Add 4:2:2 support
Instead of hardcoding only 4:2:0 support, set it dynamically
based on the first decoded frame's colorspace. It is assumed
to be the same for the whole stream.

Also, re-order the way things are set in d2vCreate, so that
everything but the chroma subsampling format is set before
we decode our test frame, so it has everything it needs to
run our custom get_buffer.

Signed-off-by: Derek Buitenhuis <derek.buitenhuis@gmail.com>